+++ packages/designer/src/framework/index.ts
@@ -4,3 +4,4 @@ export * from './assets';
 export * from './simulator';
 export * from './renderer';
 export * from './designer';
+export * from './devtools';