+++ packages/materials/CHANGELOG.md
@@ -3,6 +3,150 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.146](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.8.145...@vtj/materials@0.8.146) (2024-09-20)
+
+**Note:** Version bump only for package @vtj/materials
+
+
+
+
+
+
+## [0.8.145](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.8.144...@vtj/materials@0.8.145) (2024-09-20)
+
+**Note:** Version bump only for package @vtj/materials
+
+
+
+
+
+
+## [0.8.144](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.8.143...@vtj/materials@0.8.144) (2024-09-19)
+
+**Note:** Version bump only for package @vtj/materials
+
+
+
+
+
+## [0.8.143](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.8.142...@vtj/materials@0.8.143) (2024-09-19)
+
+**Note:** Version bump only for package @vtj/materials
+
+
+
+
+
+## [0.8.142](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.8.141...@vtj/materials@0.8.142) (2024-09-18)
+
+**Note:** Version bump only for package @vtj/materials
+
+
+
+
+
+## [0.8.141](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.8.140...@vtj/materials@0.8.141) (2024-09-18)
+
+
+### Features
+
+* ✨ 打包移除vxe-table ([27e9fd5](https://gitee.com/newgateway/vtj/commits/27e9fd501d76d2890e656f9676e07150bbd1f72f))
+
+
+
+
+
+
+## [0.8.140](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.8.139...@vtj/materials@0.8.140) (2024-09-18)
+
+**Note:** Version bump only for package @vtj/materials
+
+
+
+
+
+## [0.8.139](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.8.138...@vtj/materials@0.8.139) (2024-09-16)
+
+**Note:** Version bump only for package @vtj/materials
+
+
+
+
+
+## [0.8.138](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.8.137...@vtj/materials@0.8.138) (2024-09-16)
+
+**Note:** Version bump only for package @vtj/materials
+
+
+
+
+
+## [0.8.137](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.8.136...@vtj/materials@0.8.137) (2024-09-16)
+
+
+### Bug Fixes
+
+* 🐛 升级vue3.5 ([3b12449](https://gitee.com/newgateway/vtj/commits/3b12449447692487882539d43210d57dcc97a48a))
+
+
+
+
+
+
+## [0.8.136](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.8.135...@vtj/materials@0.8.136) (2024-09-14)
+
+**Note:** Version bump only for package @vtj/materials
+
+
+
+
+
+
+## [0.8.135](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.8.134...@vtj/materials@0.8.135) (2024-09-13)
+
+**Note:** Version bump only for package @vtj/materials
+
+
+
+
+
+## [0.8.134](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.8.133...@vtj/materials@0.8.134) (2024-09-12)
+
+**Note:** Version bump only for package @vtj/materials
+
+
+
+
+
+
+## [0.8.133](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.8.132...@vtj/materials@0.8.133) (2024-09-12)
+
+**Note:** Version bump only for package @vtj/materials
+
+
+
+
+
+## [0.8.132](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.8.131...@vtj/materials@0.8.132) (2024-09-12)
+
+**Note:** Version bump only for package @vtj/materials
+
+
+
+
+
+## [0.8.131](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.8.130...@vtj/materials@0.8.131) (2024-09-12)
+
+
+### Features
+
+* ✨ devtools module ([18e2949](https://gitee.com/newgateway/vtj/commits/18e294909e0533119f2d6f7d9fa33d470f3a6abb))
+
+
+
+
+
+
 ## [0.8.130](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.8.129...@vtj/materials@0.8.130) (2024-09-10)
 
 **Note:** Version bump only for package @vtj/materials