+++ packages/node/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.12](https://gitee.com/newgateway/vtj/compare/@vtj/node@0.8.11...@vtj/node@0.8.12) (2024-09-14)
+
+**Note:** Version bump only for package @vtj/node
+
+
+
+
+
+
 ## [0.8.11](https://gitee.com/newgateway/vtj/compare/@vtj/node@0.8.10...@vtj/node@0.8.11) (2024-09-06)