+++ platforms/web/CHANGELOG.md
@@ -3,6 +3,144 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.146](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.145...@vtj/web@0.8.146) (2024-09-20)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+
+## [0.8.145](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.144...@vtj/web@0.8.145) (2024-09-20)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+
+## [0.8.144](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.143...@vtj/web@0.8.144) (2024-09-19)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+## [0.8.143](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.142...@vtj/web@0.8.143) (2024-09-19)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+## [0.8.142](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.141...@vtj/web@0.8.142) (2024-09-18)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+## [0.8.141](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.140...@vtj/web@0.8.141) (2024-09-18)
+
+
+### Features
+
+* ✨ 打包移除vxe-table ([27e9fd5](https://gitee.com/newgateway/vtj/commits/27e9fd501d76d2890e656f9676e07150bbd1f72f))
+
+
+
+
+
+
+## [0.8.140](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.139...@vtj/web@0.8.140) (2024-09-18)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+## [0.8.139](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.138...@vtj/web@0.8.139) (2024-09-16)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+## [0.8.138](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.137...@vtj/web@0.8.138) (2024-09-16)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+## [0.8.137](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.136...@vtj/web@0.8.137) (2024-09-16)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+
+## [0.8.136](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.135...@vtj/web@0.8.136) (2024-09-14)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+
+## [0.8.135](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.134...@vtj/web@0.8.135) (2024-09-13)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+## [0.8.134](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.133...@vtj/web@0.8.134) (2024-09-12)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+
+## [0.8.133](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.132...@vtj/web@0.8.133) (2024-09-12)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+## [0.8.132](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.131...@vtj/web@0.8.132) (2024-09-12)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+## [0.8.131](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.130...@vtj/web@0.8.131) (2024-09-12)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+
 ## [0.8.130](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.129...@vtj/web@0.8.130) (2024-09-10)
 
 **Note:** Version bump only for package @vtj/web